yogi/docker-requirements.txt
requests
nvsmi
alembic==1.2.0
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
celery==4.4.7
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-API==2.0
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.1
imageio==2.5.0
imageio-ffmpeg==0.4.1
imgaug==0.3.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
opencv-contrib-python==3.4.9.33
opencv-python-headless==4.1.1.26 
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow>=6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.2
psycopg2-binary
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
python-editor==1.0.4
PyWavelets==1.0.3
pyzmq==18.1.0
qtconsole==4.5.5
seaborn==0.11.0
scikit-image==0.15.0
scikit-learn==0.23.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.8
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
